Move ManifestEndpoint to its own file
